refactor: move createStyles to constructors folder
